V4.2.1 -search bug fixed -scaling system integration -started extraction of order type functions to separate function file

+++ static/images/draw.py
@@ -1,5 +1,7 @@
 import os
 import json
+import sys
+
 from PIL import Image, ImageDraw, ImageFont
 from configs import mongo
 
@@ -25,16 +27,20 @@ def shape_plot(positions, file_name, text=[], enable_text_plot=True):
             draw.rectangle(bbox, fill="white")
             draw.text(position, str(text[i]), fill="black", font=ImageFont.truetype("c:\\server\\fonts\\upheavtt.ttf", font_size))
             text_pos.append(position)
-    im.save(file_name)
     im.show(file_name)
     if config['update_file']:
         with open('C:\\Server\\lists\\shapes.json', 'r', encoding='utf-8') as shapes_json:
             shapes = json.load(shapes_json)
         edges = len(positions) - 1
         ang = get_ang(positions)
+        if shape in shapes:
+            inp = input('Shape already exist. Overwrite? (Y/N)\n')
+            if inp != 'Y':
+                sys.exit()
+        im.save(file_name)
         shapes[shape] = {'description': '---', 'edges': edges, 'positions': text_pos,
                          'draw_positions': positions, 'ang': ang}
-        print(list(shapes.keys()))
+        # print(list(shapes.keys()))
         with open('C:\\Server\\lists\\shapes.json', 'w', encoding='utf-8') as shapes_json:
             json.dump(shapes, shapes_json)
 
@@ -43,7 +49,7 @@ def get_ang(shape_data):
     ang = []
     for ind in list(range(1, len(shape_data)-1)):
         pos1, pos2, pos3 = shape_data[ind - 1], shape_data[ind], shape_data[ind+1]
-        print(pos1, pos2, pos3)
+        # print(pos1, pos2, pos3)
         if (pos1[0] - pos2[0] == 0 or pos1[1] - pos2[1] == 0) and (pos3[0] - pos2[0] == 0 or pos3[1] - pos2[1] == 0):
             ang.append(90)
         else:
@@ -70,7 +76,7 @@ if __name__ == '__main__':
     # INPUT
     #[  185, 20  ],  [ 60,  20 ], [60, 55 ],[ 140, 55 ], [ 140, 5 ],[15, 5]]
     # print_keys()
-    pos = [(185,25), (185,5),(160,5),(120,15),(50,15),(30,5),(15,5),(15,50),(30,50),(50,25),(120,25),(150,50),(175,50),(175,20),(140,20)]
+    pos =[(15,30),(15, 50),(50,50),(140,5),(185, 5)]
     # [(180,5),(20,5), (20,50), (180,50)]
-    name = '206'
+    name = '1'
     shape_plot(pos, os.getcwd() + '\\shapes\\' + name + '.png')
